Change incorrect char.GetNumericValue references
(int)char.GetNumericValue gets the numeric equivalent of a character...i.e. '1' becomes the integer 1 and 'a' or any other letter becomes -1. What you want instead is just the char value itself, which already has an ASCII integer associated with it....i.e. 'a' is 97, 'b' is 98, etc. 

The way you have this coded now, (int)char.GetNumericValue always returns -1 and the code makes no sense in Solution 1.